SolarScan_v2/requirements.txt
torch==2.9.0
torchvision==0.24.0
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
Pillow==10.4.0
mlflow==2.16.0
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
numpy==1.26.4
